# trading-service/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Trading Service API",
        "title": "Trading Service",
        "contact": {},
        "version": "1.0"
    },
    "paths": {
...
        "/stocks": {
            "get": {
                "description": "Vraća listu svih listinga koji predstavljaju akcije.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Stocks"
                ],
                "summary": "Preuzimanje svih akcija",
                "responses": {
                    "200": {
                        "description": "Lista svih akcija",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/types.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/definitions/types.Listing"
                                            }
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "500": {
                        "description": "Interna greška servera pri preuzimanju akcija",
                        "schema": {
                            "$ref": "#/definitions/types.Response"
                        }
                    }
                }
            }
        },
        "/stocks/{ticker}": {
            "get": {
                "description": "Vraća osnovne podatke (Listing) i detalje (Stock) za akciju specificiranu tikerom.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Stocks"
                ],
                "summary": "Preuzimanje akcije po tikeru",
                "parameters": [
                    {
                        "type": "string",
                        "example": "AAPL",
                        "description": "Tiker (simbol) akcije",
                        "name": "ticker",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Detalji akcije",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/types.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "type": "object",
                                            "properties": {
                                                "details": {
                                                    "$ref": "#/definitions/types.Stock"
                                                },
                                                "listing": {
                                                    "$ref": "#/definitions/types.Listing"
                                                }
                                            }
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "404": {
                        "description": "Akcija sa datim tikerom nije pronađena",
                        "schema": {
                            "$ref": "#/definitions/types.Response"
                        }
                    },
                    "500": {
                        "description": "Interna greška servera pri preuzimanju detalja akcije",
                        "schema": {
                            "$ref": "#/definitions/types.Response"
                        }
                    }
                }
            }
        },
        "/stocks/{ticker}/history": {
            "get": {
                "description": "Preuzima dnevne istorijske podatke za akciju u vremenskom opsegu.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Stocks"
                ],
                "summary": "Istorijski podaci za akciju u vremenskom opsegu",
                "parameters": [
                    {
                        "type": "string",
                        "example": "AAPL",
                        "description": "Tiker (simbol) akcije",
                        "name": "ticker",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "format": "date",
                        "example": "2023-09-27",
                        "description": "Početni datum (YYYY-MM-DD). Podrazumevano pre 30 dana.",
                        "name": "startDate",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "format": "date",
                        "example": "2023-10-27",
                        "description": "Krajnji datum (YYYY-MM-DD). Podrazumevano današnji dan.",
                        "name": "endDate",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Lista istorijskih podataka",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/types.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/definitions/types.ListingDailyPriceInfo"
                                            }
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "Neispravan format datuma",
                        "schema": {
                            "$ref": "#/definitions/types.Response"
                        }
                    },
                    "404": {
                        "description": "Akcija sa datim tikerom nije pronađena",
                        "schema": {
                            "$ref": "#/definitions/types.Response"
                        }
                    },
                    "500": {
                        "description": "Greška pri preuzimanju istorijskih podataka iz baze",
                        "schema": {
                            "$ref": "#/definitions/types.Response"
                        }
                    }
                }
            }
        },
        "/stocks/{ticker}/history/first": {
            "get": {
                "description": "Preuzima najstariji dostupan dnevni istorijski podatak za akciju.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Stocks"
                ],
                "summary": "Prvi istorijski podatak za akciju",
                "parameters": [
                    {
                        "type": "string",
                        "example": "AAPL",
                        "description": "Tiker (simbol) akcije",
                        "name": "ticker",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Najstariji istorijski podatak",
                        "schema": {
                            "$ref": "#/definitions/types.Response"
                        }
                    },
                    "404": {
                        "description": "Akcija sa datim tikerom nije pronađena",
                        "schema": {
                            "$ref": "#/definitions/types.Response"
                        }
                    },
                    "500": {
                        "description": "Greška pri preuzimanju istorijskih podataka",
                        "schema": {
                            "$ref": "#/definitions/types.Response"
                        }
                    }
                }
            }
        },
        "/stocks/{ticker}/history/{date}": {
            "get": {
                "description": "Preuzima dnevni istorijski podatak za akciju za određeni datum.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Stocks"
                ],
                "summary": "Istorijski podatak za akciju za određeni datum",
                "parameters": [
                    {
                        "type": "string",
                        "example": "AAPL",
                        "description": "Tiker (simbol) akcije",
                        "name": "ticker",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "format": "date",
                        "example": "2023-10-26",
                        "description": "Datum u formatu YYYY-MM-DD",
                        "name": "date",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Istorijski podatak za dati datum",
                        "schema": {
                            "$ref": "#/definitions/types.Response"
                        }
                    },
                    "400": {
                        "description": "Neispravan format datuma",
                        "schema": {
                            "$ref": "#/definitions/types.Response"
                        }
                    },
                    "404": {
                        "description": "Akcija sa datim tikerom nije pronađena",
                        "schema": {
                            "$ref": "#/definitions/types.Response"
                        }
                    },
                    "500": {
                        "description": "Greška pri preuzimanju istorijskih podataka",
                        "schema": {
                            "$ref": "#/definitions/types.Response"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "types.Exchange": {
            "type": "object",
            "properties": {
                "acronym": {
                    "type": "string"
                },
                "close_time": {
                    "type": "string"
                },
                "country": {
                    "type": "string"
                },
                "currency": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "mic_code": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "open_time": {
                    "type": "string"
                },
                "timezone": {
                    "type": "string"
                }
            }
        },
...
        "types.Listing": {
            "type": "object",
            "properties": {
                "Subtype": {
                    "description": "\"ETF\", \"Common Stock\", \"Call Option\", \"Put Option\"",
                    "type": "string"
                },
                "ask": {
                    "type": "number"
                },
                "bid": {
                    "type": "number"
                },
                "contract_size": {
                    "type": "integer"
                },
                "created_at": {
                    "type": "string"
                },
                "exchange": {
                    "$ref": "#/definitions/types.Exchange"
                },
                "exchange_id": {
                    "type": "integer"
                },
                "id": {
                    "type": "integer"
                },
                "lastPrice": {
                    "type": "number"
                },
                "last_refresh": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "ticker": {
                    "type": "string"
                },
                "type": {
                    "description": "\"Stock\", \"Forex\", \"Future\", \"Option\"",
                    "type": "string"
                },
                "updated_at": {
                    "type": "string"
                }
            }
        },
        "types.ListingDailyPriceInfo": {
            "type": "object",
            "properties": {
                "availableQuantity": {
                    "type": "integer"
                },
                "change": {
                    "type": "number"
                },
                "date": {
                    "type": "string"
                },
                "high": {
                    "type": "number"
                },
                "id": {
                    "type": "integer"
                },
                "listing": {
                    "$ref": "#/definitions/types.Listing"
                },
                "listing_id": {
                    "type": "integer"
                },
                "low": {
                    "type": "number"
                },
                "price": {
                    "type": "number"
                }
            }
        },
...
        "types.Response": {
            "type": "object",
            "properties": {
                "data": {},
                "error": {
                    "type": "string"
                },
                "success": {
                    "type": "boolean"
                }
            }
        },
...
        "types.Stock": {
            "type": "object",
            "properties": {
                "dividend_yield": {
                    "type": "number"
                },
                "id": {
                    "type": "integer"
                },
                "listing": {
                    "$ref": "#/definitions/types.Listing"
                },
                "listing_id": {
                    "type": "integer"
                },
                "outstanding_shares": {
                    "type": "integer"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "BearerAuth": {
            "description": "Unesite token. Primer: \"Bearer eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9...\"",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}